removing cache from domain; adding cache to clients;
