+++ packages/client/src/participants/errors.ts
