+++ packages/client/src/participants/participantclient.ts
@@ -28,54 +28,56 @@
 
 "use strict";
 
-import nock from "nock";
 import {ILogger} from "@mojaloop/logging-bc-public-types-lib";
+import axios, {AxiosInstance, AxiosResponse, AxiosError} from "axios";
+import {
+	UnableToGetParticipantError,
+} from "./errors";
+import { IParticipant } from "@mojaloop/account-lookup-bc-domain";
+import { ILocalCache, LocalCache } from "@mojaloop/account-lookup-bc-infrastructure";
 
-export class ParticipantHttpServiceMock {
+export class ParticipantClient {
 	// Properties received through the constructor.
 	private readonly logger: ILogger;
-	private readonly BASE_URL: string;
 	// Other properties.
-	public static readonly NON_EXISTENT_PARTICIPANT_PARTY_ID: string = "non-existing-participant-id";
+	private readonly httpClient: AxiosInstance;
+	private readonly UNABLE_TO_REACH_SERVER_ERROR_MESSAGE: string = "unable to reach server";
+	private readonly _localCache: ILocalCache;
 
 	constructor(
 		logger: ILogger,
-		baseUrl: string
+		baseUrlHttpService?: string,
+		timeoutMs?: number
 	) {
 		this.logger = logger;
-		this.BASE_URL = baseUrl;
 
-		this.setUp();
+		this.httpClient = axios.create({
+			baseURL: baseUrlHttpService,
+			timeout: timeoutMs
+		});
+
+		this._localCache = new LocalCache(logger);
 	}
 
-	private setUp(): void {
-		// Get participant.
-		nock(this.BASE_URL)
-			.persist()
-			.get("/participants")
-			.query({ id: ParticipantHttpServiceMock.NON_EXISTENT_PARTICIPANT_PARTY_ID })
-			.reply(
-				(_, requestBody: any) => {
-					if (requestBody.id === ParticipantHttpServiceMock.NON_EXISTENT_PARTICIPANT_PARTY_ID) {
-						this.logger.error(`participant ${ParticipantHttpServiceMock.NON_EXISTENT_PARTICIPANT_PARTY_ID} doesn't exist`);
-						return [
-							404,
-							{message: `participant ${ParticipantHttpServiceMock.NON_EXISTENT_PARTICIPANT_PARTY_ID} doesn't exist`}
-						];
-					}
-					return [
-						404,
-						{partyId: requestBody.id}
-					];
+	async getParticipantInfo(fspId: string): Promise<IParticipant> {
+		const result = this._localCache.get("getParticipantInfo", fspId) as IParticipant;
+		
+		if (result) {
+			return result;
+		}
+		
+		try {
+			const axiosResponse: AxiosResponse = await this.httpClient.get("/participants", { params: { fspId: fspId } });
+			return axiosResponse.data;
+		} catch (e: unknown) {
+			if (axios.isAxiosError(e)) {
+				const axiosError: AxiosError = e as AxiosError;
+				if (axiosError.response !== undefined) {
+					throw new UnableToGetParticipantError((axiosError.response.data as any).message);
 				}
-			);
-	}
-
-	public disable(): void {
-		nock.restore();
-	}
-
-	public enable(): void {
-		nock.activate();
+				throw new UnableToGetParticipantError(this.UNABLE_TO_REACH_SERVER_ERROR_MESSAGE);
+			}
+			throw new UnableToGetParticipantError((e as any)?.message);
+		}
 	}
 }
