+++ packages/domain/src/aggregate.ts
@@ -44,31 +44,27 @@
 import {ILogger} from "@mojaloop/logging-bc-public-types-lib";
 import { IMessageProducer } from "@mojaloop/platform-shared-lib-messaging-types-lib";
 import { GetParticipantError, GetPartyError, NoSuchParticipantError, NoSuchParticipantFspIdError, NoSuchPartyError, UnableToAssociateParticipantError, UnableToAssociatePartyError, UnableToDisassociateParticipantError, UnableToDisassociatePartyError, UnableToGetOracleError, UnableToGetOracleProviderError } from "./errors";
-import { ILocalCache, IOracleFinder, IOracleProvider, IParticipantService} from "./interfaces/infrastructure";
-import { IParticipant, IParty, LocalCacheKeyPrefix } from "./types";
+import { IOracleFinder, IOracleProvider, IParticipantService} from "./interfaces/infrastructure";
+import { IParticipant, IParty } from "./types";
 
 export class AccountLookupAggregate {
 	private readonly _logger: ILogger;
 	private readonly _oracleFinder: IOracleFinder;
 	private readonly _oracleProviders: IOracleProvider[];
     private readonly _messageProducer: IMessageProducer;
-    private readonly _localCache: ILocalCache;
     private readonly _participantService: IParticipantService;
-    private participantCacheKeyPrefix: LocalCacheKeyPrefix = "participant";  
 
 	constructor(
 		logger: ILogger,
         oracleFinder:IOracleFinder,
         oracleProviders:IOracleProvider[],
         messagePublisher:IMessageProducer,
-        localCache:ILocalCache,
         participantService: IParticipantService
 	) {
 		this._logger = logger;
 		this._oracleFinder = oracleFinder;
 		this._oracleProviders = oracleProviders;
         this._messageProducer = messagePublisher;
-        this._localCache = localCache;
         this._participantService = participantService;
     }
 
@@ -215,14 +211,6 @@ export class AccountLookupAggregate {
     //Participant.
     async getParticipantByTypeAndId(participantType:string, participantId:string):Promise<IParticipant|null|undefined>{
         let participant: IParticipant | null; 
-       
-        participant = this.extractParticipantFromCache(participantType, participantId);
-        
-        // Return a cached response
-        if(participant) {
-            return participant;
-        }
-
         
         const oracleProvider = await this.getOracleProvider(participantType);
         
@@ -243,8 +231,6 @@ export class AccountLookupAggregate {
             this._logger.debug(`No participant by type: ${participantType} and id: ${participantId} found`);
             throw new NoSuchParticipantError();
         }
-
-        this.storeParticipantInCache(participant)
         
         return participant;
 
@@ -252,14 +238,6 @@ export class AccountLookupAggregate {
 
     async getParticipantByTypeAndIdAndSubId(participantType:string, participantId:string, participantSubId:string):Promise<IParticipant|null|undefined>{
         let participant: IParticipant | null; 
-       
-        participant = this.extractParticipantFromCache(participantType, participantId, participantSubId);
-        
-        // Return a cached response
-        if(participant) {
-            return participant;
-        }
-
         
         const oracleProvider = await this.getOracleProvider(participantType);
         
@@ -281,7 +259,6 @@ export class AccountLookupAggregate {
             throw new NoSuchParticipantError();
         }
 
-        this.storeParticipantInCache(participant)
         
         return participant;
 
@@ -345,17 +322,4 @@ export class AccountLookupAggregate {
 		return oracleProvider;
 	}
     
-    private extractParticipantFromCache(participantType: string, participantId: string, participantSubId?:string):IParticipant|null {
-        const cachedParticipant = this._localCache.get(this.participantCacheKeyPrefix, participantType, participantId, participantSubId ?? "") as IParticipant;
-        return cachedParticipant;
-    }
-
-    private storeParticipantInCache(participant:IParticipant):void {
-        try{
-            this._localCache.set(participant, this.participantCacheKeyPrefix, participant.type, participant.id, participant.subId ?? "");
-        }
-        catch(error){
-            this._logger.error("Unable to store participant in cache " + error);
-        }
-    }
 }
